+++ src/components/radio-button/radio-button.test.ts
@@ -89,8 +89,9 @@ describe('<sl-radio-button>', () => {
 
       form.addEventListener('submit', submitHandler);
       radio.click();
-      button.click();
+      await radio.updateComplete;
 
+      button.click();
       await waitUntil(() => submitHandler.calledOnce);
 
       expect(formData!.get('a')).to.equal('2');