+++ src/components/radio/radio.test.ts
@@ -90,8 +90,9 @@ describe('<sl-radio>', () => {
 
       form.addEventListener('submit', submitHandler);
       radio.click();
-      button.click();
+      await radio.updateComplete;
 
+      button.click();
       await waitUntil(() => submitHandler.calledOnce);
 
       expect(formData!.get('a')).to.equal('2');